API

Projects

ResearchProjects.AbstractProject — Type

abstract type AbstractProject{FT}

Hierarchy of AbstractProject:

ClumpingFactor2020

FluxTower2020

GriddingMachine2020

LeafInvestment2020

LeafInvestment2021

NocturnalGS2020

OptimalGM2020

SIFComparison2020

ResearchProjects.ClumpingFactor2020 — Type

struct ClumpingFactor2020{FT}

ResearchProjects.FluxTower2020 — Type

struct FluxTower2020{FT}

ResearchProjects.GriddingMachine2020 — Type

struct GriddingMachine2020{FT}

ResearchProjects.LeafInvestment2020 — Type

struct LeafInvestment2020{FT}

ResearchProjects.LeafInvestment2021 — Type

struct LeafInvestment2021{FT}

ResearchProjects.NocturnalGS2020 — Type

struct NocturnalGS2020{FT}

ResearchProjects.OptimalGM2020 — Type

struct OptimalGM2020{FT}

ResearchProjects.SIFComparison2020 — Type

struct SIFComparison2020{FT}

General functions

ResearchProjects.create_spac — Function

create_spac(proj::FluxTower2020{FT},
            sm::AbstractStomatalModel) where {FT<:AbstractFloat}
create_spac(proj::LeafInvestment2020{FT},
            lat::FT,
            ele::FT;
            spp::String) where {FT<:AbstractFloat}
create_spac(proj::LeafInvestment2020{FT},
            df::DataFrame,
            lat::FT,
            ele::FT;
            spp::String
            gmax::FT) where {FT<:AbstractFloat}

Create SPACSimple or SPACMono struct , given

proj LeafInvestment2020 or FluxTower2020 type project

sm AbstractStomatalModel type stomatal model. Default is OSMWang

lat Latitude. Optional, default at Flagstaff

ele Elevation. Optional, default at Flagstaff

spp AbstractSpecies type identifier. Optional, default is "aspen"

df Preloaded dataframe

gmax Maximal stomatal conductance. Optional, default is 0.1

ResearchProjects.dynamic_workers — Function

dynamic_workers(nTH::Int)
dynamic_workers(project::FluxTower2020{FT},
                nTH::Int) where {FT<:AbstractFloat}
dynamic_workers(project::LeafInvestment2020{FT},
                nTH::Int) where {FT<:AbstractFloat}

Change the number of workers according to CPU_THREADS, given

nTH Number of thread required

project FluxTower2020 or LeafInvestment2020 type project identifier

Species indentifier

ResearchProjects.AbstractSpecies — Type

abstract type AbstractSpecies

Hierarchy of AbstractSpecies

PopulusTremuloidesSperry2019

ResearchProjects.NiwotRidgeWang2020 — Type

struct NiwotRidgeWang2020

ResearchProjects.PopulusTremuloidesSperry2019 — Type

struct PopulusTremuloidesSperry2019

Data from Sperry et al., (2019). The impact of rising CO2 and acclimation on the response of US forests to global warming.

Ptoject ClumpingIndex2020

ResearchProjects.fit_leafbio — Function

fit_leafbio(proj::ClumpingFactor2020{FT},
            refs::Array{FT,1},
            use_latex::Bool) where {FT<:AbstractFloat}

Fit leaf bio parameters to match given spectrum data, given

proj ClumpingFactor2020 type project control

refs Array of reference spectrum data

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.spectrum_diff! — Function

spectrum_diff!(
            proj::ClumpingFactor2020{FT},
            wls::WaveLengths{FT},
            leaf::LeafBios{FT},
            vars::Array{FT,1},
            refs::Array{FT,1},
            sims::Array{FT,1}) where {FT<:AbstractFloat}

Calculate the sum of square of difference between providing spectrum values and simulated spectrum values, given

proj ClumpingFactor2020 type project control

wls WaveLengths type struct from CanopyLayers module

leaf LeafBio type struct from CanopyLayers module

vars Array of variables to fit

refs Array of reference spectrum data

sims Array of simulated spectrum data (updated within function)

Project FluxTower2020

ResearchProjects.calculate_fitting_density — Function

calculate_fitting_density(
            proj::FluxTower2020{FT};
            nTH::Int = 20,
            fitGmax::Bool = true,
            perday::Bool = true
) where {FT<:AbstractFloat}

Calculate the scatter density of fitted simulations, given

proj FluxTower2020 type project identifier

nTH Number of threads to run the code in parallel

fitGmax If true, Gmax has been fitted

perday Optional. If true, seperate the data per day

ResearchProjects.diurnal_simulation! — Function

diurnal_simulation!(
            proj::FluxTower2020{FT},
            node::SPACMono{FT},
            df::DataFrame,
            rbase::Q10TD) where {FT<:AbstractFloat}

Run diurnal cycle of land model simulation, given

proj FluxTower2020 type project identifier

node SPACMono type struct from SoilPlantAirContinuum module

df Dataframe of weather data. Note that simulated data are stored in df.

rbase Q10 based temperature dependency struct

ResearchProjects.doy_to_int — Function

doy_to_int(year::Int, doy::Int)

Convert date to int number to work with flux tower datasets, given

year Year with the format of XXXX

doy Day of year (typically 1-365, 1-366 for leap years)

ResearchProjects.fit_flux_tower_data — Function

fit_flux_tower_data(
            proj::FluxTower2020{FT},
            params::Array,
            label::String;
            nTH::Int = 18,
            fitGmax::Bool = true
) where {FT<:AbstractFloat}
fit_flux_tower_data(
            proj::FluxTower2020{FT},
            sm::ESMBallBerry{FT},
            params::Array,
            label::String;
            nTH::Int = 18,
            fitGmax::Bool = true
) where {FT<:AbstractFloat}
fit_flux_tower_data(
            proj::FluxTower2020{FT},
            sm::ESMMedlyn{FT},
            params::Array,
            label::String;
            nTH::Int = 18,
            fitGmax::Bool = true
) where {FT<:AbstractFloat}
fit_flux_tower_data(
            proj::FluxTower2020{FT};
            nTH::Int = 18,
            fitGmax::Bool = true,
            perday::Bool = true
) where {FT<:AbstractFloat}

Fit Vcmax, Rbase, Gmin, and Gmax (and g0 and g1 as well if sm is given), given

proj FluxTower2020 type project identifier

params Array of parameters to pass into threadings

nTH How many threads to run the code in parallel

fitGmax If true, fit Gmax as well

sm ESMBallBerry or ESMMedlyn type stomatal model

perday Optional. If true, seperate the data per day

ResearchProjects.fit_vr! — Function

fit_vr!(param::Array)

Fit Vcmax and Rbase, given

param An array of parameters to pass into threaded workers, the param consists of project identifier, SPACMono type struct, dataframe to operate, floating number type, year identifier, and doy identifier

ResearchProjects.fit_vrg! — Function

fit_vrg!(param::Array)

Fit Vcmax, Rbase, and Gmax, given

param An array of parameters to pass into threaded workers, the param consists of project identifier, SPACMono type struct, dataframe to operate, floating number type, year identifier, and doy identifier

ResearchProjects.fit_vrgg! — Function

fit_vrgg!(param::Array)

Fit Vcmax, Rbase, g0, and g1 (for Ball Berry Model), given

param An array of parameters to pass into threaded workers, the param consists of project identifier, SPACMono type struct, dataframe to operate, floating number type, year identifier, and doy identifier

ResearchProjects.fit_vrggg! — Function

fit_vrggg!(param::Array)

Fit Vcmax, Rbase, and Gmax, g0 and g1 (for Ball Berry Model), given

param An array of parameters to pass into threaded workers, the param consists of project identifier, SPACMono type struct, dataframe to operate, floating number type, year identifier, and doy identifier

ResearchProjects.plot_density_scatter — Function

plot_density_scatter(
            proj::FluxTower2020{FT};
            saving::Bool,
            use_latex::Bool,
            fitGmax::Bool = true,
            perday::Bool = true
) where {FT<:AbstractFloat}

Plot the density scatter of modeled fluxes vs. observed data, given

proj FluxTower2020 type project identifier

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

fitGmax If true, Gmax has been fitted

perday Optional. If true, seperate the data per day

ResearchProjects.plot_fitted_variable — Function

plot_fitted_variable(
            proj::FluxTower2020{FT},
            varname::String,
            figname::String,
            colors::Array{String,1} = ["lightgray", "darkred", "salmon"],
            nday::Int=5;
            saving::Bool = false,
            use_latex::Bool = true,
            vrggg::Bool = false,
            fitGmax::Bool = true,
            perday::Bool = true
) where {FT<:AbstractFloat}
plot_fitted_variable(
            proj::FluxTower2020{FT},
            nday::Int=5;
            saving::Bool = false,
            use_latex::Bool = true,
            fitGmax::Bool = true
) where {FT<:AbstractFloat}

Plot fitted variables, given

proj FluxTower2020 type project identifier

varname Variable name to plot

figname Figure name that display on window title bar

colors Colors for plots: individual point, mean, mean ± sd

nday Number of days to calculate running average

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

vrggg If true, use the version that fitted g0 and g1

fitGmax If true, Gmax has been fitted

perday Optional. If true, seperate the data per day

ResearchProjects.plot_model_example — Function

plot_model_example(
            proj::FluxTower2020{FT};
            saving::Bool = false,
            use_latex::Bool = true,
            fitGmax::Bool = true,
            perday::Bool = true
) where {FT<:AbstractFloat}

Plot examples of modeled fluxes vs. observed data, given

proj FluxTower2020 type project identifier

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

fitGmax If true, Gmax has been fitted

perday Optional. If true, seperate the data per day

ResearchProjects.print_stats_table — Function

print_stats_table(
            proj::FluxTower2020{FT};
            saving::Bool = false,
            perday::Bool = true
) where {FT<:AbstractFloat}

Calculate the statistics of the simulations, given

proj FluxTower2020 type project identifier

saving Optional. If true, save the table to a CSV file

perday Optional. If true, seperate the data per day

ResearchProjects.query_dataframe — Function

query_dataframe(
            proj::FluxTower2020{FT},
            year::Int
) where {FT<:AbstractFloat}
query_dataframe(
            proj::FluxTower2020{FT},
            data::DataFrame,
            year::Int,
            doy::Int
) where {FT<:AbstractFloat}
query_dataframe(
            proj::FluxTower2020{FT},
            year::Int,
            doy::Int
) where {FT<:AbstractFloat}
query_dataframe(
            proj::FluxTower2020{FT},
            years::Array{Int,1} = [2019],
            sm::AbstractStomatalModel = OSMWang{FT}();
            perday::Bool = true
) where {FT<:AbstractFloat}

Query DataFrame formated data, given

proj FluxTower2020 type project identifier

year Which year of data to query

data DataFrame of data

doy Which day of data to query

years Array of years

sm Pre-defined stomatal model

perday Optional. If true, seperate the data per day

Note that if year and doy are provided, the function returns a dataframe with data from flux tower dataset, as well as new generated columns, including ModC and MODE used to store values from land model simulation. If years and sm are provided, the function returns an array of data to pass to threading functions.

ResearchProjects.update_gsw! — Function

update_gsw!(clayer::CanopyLayer{FT},
            sm::AbstractStomatalModel{FT},
            photo_set::AbstractPhotoModelParaSet{FT},
            envir::AirLayer{FT},
            Δt::FT) where {FT<:AbstractFloat}

Update g_sw prognostically, given

clayer A canoy layer –- CanopyLayer type struct

sm AbstractStomatalModel type stomatal model

photo_set AbstractPhotoModelParaSet type photosynthesis model

envir AirLayer type environmental conditions

Δt Time interval for prognostic stomatal conductance

Ptoject GriddingMachine2020

ResearchProjects.compare_datasets — Function

compare_datasets(proj::GriddingMachine2020{FT};
            saving::Bool=false,
            use_latex::Bool=true
) where {FT<:AbstractFloat}

Compare the inconsistency among datasets, given

proj GriddingMachine2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

Project LeafInvestment2020

ResearchProjects.fit_invest_cn — Function

ResearchProjects.get_optima_for_invest — Function

get_optima_for_invest(param::Array)

Local function to optimize leaf investment in multiple threading manner, given

param Array of parameters include the following (for sensitivity analayis)

site name

sensitivity analysis variable

sensitivity analysis variable ratio

weather year

SPACSimple struct from the SPAC module

AbstractPhotoModelParaSet struct from the Photosynthesis module

Weather data as a 2D array

param Array of parameters include the following (for invest_cn contour)

progress percentage

invest_c

invest_n

SPACSimple struct from the SPAC module

AbstractPhotoModelParaSet struct from the Photosynthesis module

Weather data as a 2D array

ResearchProjects.leaf_diff — Function

ResearchProjects.mean_LVGC — Function

mean_LVGC(
            project::LeafInvestment2020{FT},
            data::DataFrame,
            site::String,
            lab::String,
            ratios::Array{FT,1},
            use_laba::Bool
) where {FT<:AbstractFloat}

Calculate and return array of mean LAGA, VMAX, GSCP, and CICA at different trait settings, given

project LeafInvestment2020 type project

data DataFrame that contains data of sensitivity analysis

site Which site to filter

lab Sensitivity to which variable

ratios Array of values for how much does the variable vary from default

use_laba Optional. If true use LABA, else use LAGA

ResearchProjects.plot_a2gs_ci2ca_comparison — Function

plot_a2gs_ci2ca_comparison(
            project::LeafInvestment2020{FT};
            saving::Bool=false
) where {FT<:AbstractFloat}

Plot the comparison of A/Gs and Ci/Ca variation to reply the reviewer's question, given

project LeafInvestment2020 type of project

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_annual_simulation — Function

plot_annual_simulation(
            project::LeafInvestment2020{FT};
            saving::Bool,
            use_latex::Bool
) where {FT<:AbstractFloat}

Plot annual simulation, given

project LeafInvestment2020 type project

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_invest_cn_contour_example — Function

plot_invest_cn_contour_example(
            project::LeafInvestment2020{FT};
            saving::Bool = false
) where {FT<:AbstractFloat}

Plot the investment map contour to fit investC and investN, given

project LeafInvestment2020 type of project

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_laba_vmax_combo — Function

plot_laba_vmax_combo(
            project::LeafInvestment2020{FT};
            saving::Bool = false
) where {FT<:AbstractFloat}

Sort and plot the relative sensitivitiesy to xylem hydraulics, leaf, water supply, and air, given

project LeafInvestment2020 type project

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_leaf_optimum_example — Function

plot_leaf_optimum_example(
            project::LeafInvestment2020{FT};
            laba_list::Array{FT,1},
            vmax_list::Array{FT,1},
            saving::Bool = false
) where {FT<:AbstractFloat}

Plot annual profit at different leaf investment, given

project LeafInvestment2020 type project

laba_list Given array of LA:BA to grid

vmax_list Given array of Vcmax to grid

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_sa_on_axes — Function

plot_sa_on_axes(
            project::LeafInvestment2020{FT},
            _axes::Array,
            labs::Array{String,1},
            rats::Array{Array{FT,1},1},
            nrow::Int;
            high_gmax::Bool,
            use_laba::Bool
) where {FT<:AbstractFloat}

Plot sensitivity analysis on axes, given

project LeafInvestment2020 type project

_axes Array of axes to plot on

labs Sensitivity to which variables

rat Array of array of ratios

nrow Row count of the figure

high_gmax Optional. If true, plot the case of high base gmax

use_laba Optional. If true use LABA, else use LAGA

ResearchProjects.plot_sensitivity_analysis — Function

plot_sensitivity_analysis(
            project::LeafInvestment2020{FT};
            saving::Bool = false
) where {FT<:AbstractFloat}

Plot the sensitivity to xylem hydraulics, leaf, water supply, and air, given

project LeafInvestment2020 type project

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_sensitivity_analysis_si — Function

plot_sensitivity_analysis_si(
            project::LeafInvestment2020{FT};
            saving::Bool = false
) where {FT<:AbstractFloat}

Plot the sensitivity to gmax at high gmax, given

project LeafInvestment2020 type project

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_sensitivity_bar — Function

plot_sensitivity_bar(
            project::LeafInvestment2020{FT};
            saving::Bool = false
) where {FT<:AbstractFloat}

Sort and plot the relative sensitivitiesy to xylem hydraulics, leaf, water supply, and air, given

project LeafInvestment2020 type project

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_sun_shade_optimum — Function

plot_sun_shade_optimum(
            project::LeafInvestment2020{FT};
            zenith::FT,
            rall::FT,
            saving::Bool
) where {FT<:AbstractFloat}

Plot the 3D contour of optimizer at different sunlit and shaded flow rates, given

project LeafInvestment2020 type of project

zenith Given zenith angle (optional)

rall Given total radiation (optional)

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.run_invest_cn_contour_example — Function

run_invest_cn_contour_example(
            project::LeafInvestment2020{FT},
            nTH::Int,
            array_c::Array{FT,1},
            array_n::Array{FT,1}
) where {FT<:AbstractFloat}

Generate investment map data in multiple threading manner, given

project LeafInvestment2020 type of project

nTH Number of threadings to run the simulation

array_c Given array of investC

array_n Given array of investN

ResearchProjects.run_sensitivity_analysis — Function

run_sensitivity_analysis(
            project::LeafInvestment2020{FT};
            nTH::Int,
            siteids::Array{Int,1},
            selection::Array{String,1}
) where {FT<:AbstractFloat}

Run sensitivity analysis for optimal leaf investment, given

project LeafInvestment2020 type project

nTH Number of threads

siteids Indices of selected sites

selection List of parameters to do sensitivity on

ResearchProjects.run_sensitivity_analysis_gmax — Function

run_sensitivity_analysis(
            project::LeafInvestment2020{FT};
            nTH::Int,
            siteids::Array{Int,1},
            selection::Array{String,1}
) where {FT<:AbstractFloat}

Run sensitivity analysis for optimal leaf investment, given

project LeafInvestment2020 type project

nTH Number of threads

siteids Indices of selected sites

selection List of parameters to do sensitivity on

Project NocturnalGS2020

ResearchProjects.plot_framework — Function

plot_framework(
            project::NocturnalGS2020{FT};
            saving::Bool
) where {FT<:AbstractFloat}

Plot model framework, given

project NocturnalGS2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_gswn_vs_pcd — Function

plot_gswn_vs_pcd(
            project::NocturnalGS2020{FT};
            saving::Bool
) where {FT<:AbstractFloat}

Plot gswn vs Psoil, Ca, and D, given

project NocturnalGS2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_gswn_vs_r — Function

plot_gswn_vs_r(
            project::NocturnalGS2020{FT};
            saving::Bool
) where {FT<:AbstractFloat}

Plot gswn vs leaf respiration rate, given

project NocturnalGS2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_gswn_vs_tem — Function

plot_gswn_vs_tem(
            project::NocturnalGS2020{FT};
            saving::Bool
) where {FT<:AbstractFloat}

Plot gswn vs temperature, given

project NocturnalGS2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_gswn_vs_time — Function

plot_gswn_vs_time(
            project::NocturnalGS2020{FT};
            saving::Bool
) where {FT<:AbstractFloat}

Plot gswn vs time, given

project NocturnalGS2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_gswn_vpd — Function

plot_gswn_vpd(
            project::NocturnalGS2020{FT};
            saving::Bool
) where {FT<:AbstractFloat}

Plot gswn vs VPD, given

project NocturnalGS2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.calculate_dAdE! — Function

calculate_dAdE!(
            leaf::Leaf{FT},
            env::AirLayer{FT},
            psm::AbstractPhotoModelParaSet{FT},
            a_n::FT,
            c_a::FT,
            e_leaf::FT,
            gbw::FT,
            p_atm::FT,
            par::FT,
            t_leaf::FT,
            vp_air::FT,
            dTdE::Bool
) where {FT<:AbstractFloat}

Calculate the marginal water use efficiency by firstly fitting the photosynthetic capacity and secondly numerically computing dAdE, given

leaf Leaf type struct that stores leaf photosynthetic parameters

env AirLayer type struct that stores environmental conditions

psm AbstractPhotoModelParaSet type struct that stores photosynthesis model parameters, like the temperature dependencies

a_n Given net photosynthetic rate

c_a Given CO₂ concentration in ppm

e_leaf Given leaf transpiration rate

gbw Given leaf boundary layer conductance for H₂O

p_atm Given atmospheric pressure

par Given photosynthetically active radiation

t_leaf Given leaf temperature

vp_air Given vapor pressure in the air

dTdE::Bool If true, account for leaf cooling from transpiration

ResearchProjects.calculate_dRdE — Function

calculate_dRdE(
            psm::AbstractPhotoModelParaSet{FT},
            r_n::FT,
            t_leaf::FT,
            wind::FT
) where {FT<:AbstractFloat}

Calculate marginal respiratory reduction, given

psm AbstractPhotoModelParaSet type struct that stores photosynthesis model parameters, like the temperature dependencies

r_n Given leaf respiration rate

t_leaf Given leaf temperature

wind Given wind speed

ResearchProjects.fitness_factor — Function

fitness_factor(
            project::NocturnalGS2020{FT};
            dTdE::Bool,
            display_result::Bool
) where {FT<:AbstractFloat}

Calculate the mean and standard deviation of the fitness factors, given

project NocturnalGS2020 type project control

dTdE Whether accounting for leaf cooling effect for daytime photosynthesis

display_result Optional. If true, display the results

Project OptimalGM2020

ResearchProjects.a_cc_vs_gm — Function

a_cc_vs_gm(psm::AbstractPhotoModelParaSet{FT},
           leaf::Leaf{FT},
           env::AirLayer{FT};
           gsc::FT) where {FT<:AbstractFloat}

Calculate chloroplast CO₂ and net photosynthetic rate at different mesophyll conductance, given

psm Photosynthesis model parameter set from Photosynthesis module

leaf Leaf struct from Photosynthesis module

env AirLayer struct from Photosynthesis module

gsc Stomatal conductance to CO₂ (default at 0.2)

ResearchProjects.a_vs_cc — Function

a_vs_cc(psm::AbstractPhotoModelParaSet{FT},
        leaf::Leaf{FT},
        env::AirLayer{FT};
        ccs::Array{FT,1}) where {FT<:AbstractFloat}

Calculate net photosynthetic rate at different chloroplast CO₂, given

psm Photosynthesis model parameter set from Photosynthesis module

leaf Leaf struct from Photosynthesis module

env AirLayer struct from Photosynthesis module

ccs Chloroplast CO₂ partial pressure array (default from 5 to 45 Pa)

ResearchProjects.optimal_gm — Function

optimal_gm(proj::OptimalGM2020{FT},
           psm::AbstractPhotoModelParaSet{FT},
           canopyi::CanopyLayer{FT},
           hs::LeafHydraulics{FT},
           env::AirLayer{FT},
           sm::OSMWang{FT},
           f_cost::FT) where {FT<:AbstractFloat}

Calculate the optimal mesophyll conductance, given

proj OptimalGM2020 type project control

psm Photosynthesis model parameter set from Photosynthesis module

canopyi CanopyLayer type struct from StomataModels module

hs LeafHydraulics type struct from PlantHydraulics module

env AirLayer type struct from Photosynthesis module

sm OSMWang type stomatal model from StomataModels module

f_cost Marginal cost in photosynthetic rate per marginal increase in mesophyll conductance

ResearchProjects.optimal_gss_gms — Function

optimal_gss_gms(proj::OptimalGM2020{FT},
                psm::AbstractPhotoModelParaSet{FT},
                can::CanopyLayer{FT},
                env::AirLayer{FT},
                hs::LeafHydraulics{FT},
                sm::OSMWang{FT},
                lab::String) where {FT<:AbstractFloat}

Calculate optimal stomatal and mesophyll conductance arrays, given

proj OptimalGM2020 type project control

psm Photosynthesis model parameter set from Photosynthesis module

leaf Leaf struct from Photosynthesis module

env AirLayer struct from Photosynthesis module

hs LeafHydraulics struct from PlantHydraulics module

sm OSMWang type of stomatal model from StomataModels module

lab Which variable to modify and run sensitivity analysis

ResearchProjects.plot_data_fittings — Function

plot_data_fittings(
            proj::OptimalGM2020{FT};
            saving::Bool = false
) where {FT<:AbstractFloat}

Plot the model predictions and real data, given

proj OptimalGM2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_gm_prediction — Function

plot_gm_prediction(proj::OptimalGM2020{FT},
                   ax,
                   lab::String;
                   offset_x::FT,
                   multip_y::FT,
                   plot_gs::Bool,
                   vcm::FT,
                   jm::FT,
                   gm_ls::String) where {FT<:AbstractFloat}

Plot mesophyll conductance response on axis, given

proj OptimalGM2020 type project control

ax Axis to plot on

lab Identifier to run sensitivity analysis

offset_x An offset on x-axis, e.g., to convert K to °C

multip_y A multiplier on y-axis

plot_gs Optional. If true, plot stomatal conductance as well

vcm Vcmax25 to compute leaf photosynthetic rate

jm Jmax25 to compute leaf photosynthetic rate

gm_ls Line style to plot mesophyll conductance

ResearchProjects.plot_gm_var_maps — Function

plot_gm_var_maps(
            proj::OptimalGM2020{FT};
            saving::Bool = false
) where {FT<:AbstractFloat}

Plot the Ci-Gm and PAR-Gm contour map of ∂A/∂Gm, given

proj OptimalGM2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_optimal_gm_framework — Function

plot_optimal_gm_framework(
            proj::OptimalGM2020{FT};
            saving::Bool
) where {FT<:AbstractFloat}

Plot model framework, given

proj OptimalGM2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_optimal_gm_predictions — Function

plot_optimal_gm_predictions(
            proj::OptimalGM2020{FT};
            saving::Bool = false) where {FT<:AbstractFloat}

Plot model predictions of how optimal mesophyll respond to the environment, given

proj OptimalGM2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_optimal_gm_vs_ca — Function

plot_optimal_gm_vs_ca(
            proj::OptimalGM2020{FT};
            saving::Bool
) where {FT<:AbstractFloat}

Plot model framework, given

proj OptimalGM2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_optimal_gm_vs_gs — Function

plot_optimal_gm_vs_gs(
            proj::OptimalGM2020{FT};
            saving::Bool
) where {FT<:AbstractFloat}

Plot model framework, given

proj OptimalGM2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_optimal_gm_vs_par — Function

plot_optimal_gm_vs_par(
            proj::OptimalGM2020{FT};
            saving::Bool
) where {FT<:AbstractFloat}

Plot model framework, given

proj OptimalGM2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.∂A∂Gm — Function

∂A∂Gm(proj::OptimalGM2020{FT},
      psm::AbstractPhotoModelParaSet{FT},
      leaf::Leaf{FT},
      env::AirLayer{FT},
      ci::FT,
      gm::FT) where {FT<:AbstractFloat}

Calculate the marginal photosynthetic gain per marginal increase in mesophyll conductance, given

proj OptimalGM2020 type project control

psm Photosynthesis model parameter set from Photosynthesis module

leaf Leaf type struct from Photosynthesis module

env AirLayer type struct from Photosynthesis module

ci Given leaf internal CO₂ in the intercellular airspace [Pa]

gm Given mesophyll condutcance

SIFComparison2020

ResearchProjects.derive_spectrum — Function

derive_spectrum(proj::SIFComparison2020{FT};
                saving::Bool,
                use_latex::Bool) where {FT<:AbstractFloat}

Derive and plot the albedo and SIF spectrum, given

proj SIFComparison2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font