API

Projects

ResearchProjects.AbstractProject — Type

abstract type AbstractProject{FT}

Hierarchy of AbstractProject:

ClumpingFactor2020

FluxTower2020

GriddingMachine2020

LeafInvestment2020

LeafInvestment2021

NocturnalGS2020

OptimalGM2020

SIFComparison2020

ResearchProjects.ClumpingFactor2020 — Type

struct ClumpingFactor2020{FT}

ResearchProjects.GriddingMachine2020 — Type

struct GriddingMachine2020{FT}

ResearchProjects.LeafInvestment2020 — Type

struct LeafInvestment2020{FT}

ResearchProjects.LeafInvestment2021 — Type

struct LeafInvestment2021{FT}

ResearchProjects.NocturnalGS2020 — Type

struct NocturnalGS2020{FT}

ResearchProjects.OptimalGM2020 — Type

struct OptimalGM2020{FT}

ResearchProjects.SIFComparison2020 — Type

struct SIFComparison2020{FT}

General functions

ResearchProjects.dynamic_workers — Function

dynamic_workers(nTH::Int)
dynamic_workers(project::FluxTower2020{FT},
                nTH::Int) where {FT<:AbstractFloat}
dynamic_workers(project::LeafInvestment2020{FT},
                nTH::Int) where {FT<:AbstractFloat}

Change the number of workers according to CPU_THREADS, given

nTH Number of thread required

project FluxTower2020 or LeafInvestment2020 type project identifier

Ptoject ClumpingIndex2020

ResearchProjects.fit_leafbio — Function

fit_leafbio(proj::ClumpingFactor2020{FT},
            refs::Array{FT,1},
            use_latex::Bool) where {FT<:AbstractFloat}

Fit leaf bio parameters to match given spectrum data, given

proj ClumpingFactor2020 type project control

refs Array of reference spectrum data

use_latex Optional. If true, use latex and serif font

Ptoject GriddingMachine2020

ResearchProjects.compare_datasets — Function

compare_datasets(proj::GriddingMachine2020{FT};
            saving::Bool=false,
            use_latex::Bool=true
) where {FT<:AbstractFloat}

Compare the inconsistency among datasets, given

proj GriddingMachine2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

Project LeafInvestment2020

ResearchProjects.fit_invest_cn — Function

ResearchProjects.plot_a2gs_ci2ca_comparison — Function

plot_a2gs_ci2ca_comparison(
            project::LeafInvestment2020{FT};
            saving::Bool=false
) where {FT<:AbstractFloat}

Plot the comparison of A/Gs and Ci/Ca variation to reply the reviewer's question, given

project LeafInvestment2020 type of project

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_annual_simulation — Function

plot_annual_simulation(
            project::LeafInvestment2020{FT};
            saving::Bool,
            use_latex::Bool
) where {FT<:AbstractFloat}

Plot annual simulation, given

project LeafInvestment2020 type project

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_invest_cn_contour_example — Function

plot_invest_cn_contour_example(
            project::LeafInvestment2020{FT};
            saving::Bool = false
) where {FT<:AbstractFloat}

Plot the investment map contour to fit investC and investN, given

project LeafInvestment2020 type of project

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_laba_vmax_combo — Function

plot_laba_vmax_combo(
            project::LeafInvestment2020{FT};
            saving::Bool = false
) where {FT<:AbstractFloat}

Sort and plot the relative sensitivitiesy to xylem hydraulics, leaf, water supply, and air, given

project LeafInvestment2020 type project

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_leaf_optimum_example — Function

plot_leaf_optimum_example(
            project::LeafInvestment2020{FT};
            laba_list::Array{FT,1},
            vmax_list::Array{FT,1},
            saving::Bool = false
) where {FT<:AbstractFloat}

Plot annual profit at different leaf investment, given

project LeafInvestment2020 type project

laba_list Given array of LA:BA to grid

vmax_list Given array of Vcmax to grid

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_sensitivity_analysis — Function

plot_sensitivity_analysis(
            project::LeafInvestment2020{FT};
            saving::Bool = false
) where {FT<:AbstractFloat}

Plot the sensitivity to xylem hydraulics, leaf, water supply, and air, given

project LeafInvestment2020 type project

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_sensitivity_analysis_si — Function

plot_sensitivity_analysis_si(
            project::LeafInvestment2020{FT};
            saving::Bool = false
) where {FT<:AbstractFloat}

Plot the sensitivity to gmax at high gmax, given

project LeafInvestment2020 type project

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_sensitivity_bar — Function

plot_sensitivity_bar(
            project::LeafInvestment2020{FT};
            saving::Bool = false
) where {FT<:AbstractFloat}

Sort and plot the relative sensitivitiesy to xylem hydraulics, leaf, water supply, and air, given

project LeafInvestment2020 type project

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_sun_shade_optimum — Function

plot_sun_shade_optimum(
            project::LeafInvestment2020{FT};
            zenith::FT,
            rall::FT,
            saving::Bool
) where {FT<:AbstractFloat}

Plot the 3D contour of optimizer at different sunlit and shaded flow rates, given

project LeafInvestment2020 type of project

zenith Given zenith angle (optional)

rall Given total radiation (optional)

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.run_invest_cn_contour_example — Function

run_invest_cn_contour_example(
            project::LeafInvestment2020{FT},
            nTH::Int,
            array_c::Array{FT,1},
            array_n::Array{FT,1}
) where {FT<:AbstractFloat}

Generate investment map data in multiple threading manner, given

project LeafInvestment2020 type of project

nTH Number of threadings to run the simulation

array_c Given array of investC

array_n Given array of investN

ResearchProjects.run_sensitivity_analysis — Function

run_sensitivity_analysis(
            project::LeafInvestment2020{FT};
            nTH::Int,
            siteids::Array{Int,1},
            selection::Array{String,1}
) where {FT<:AbstractFloat}

Run sensitivity analysis for optimal leaf investment, given

project LeafInvestment2020 type project

nTH Number of threads

siteids Indices of selected sites

selection List of parameters to do sensitivity on

ResearchProjects.run_sensitivity_analysis_gmax — Function

run_sensitivity_analysis(
            project::LeafInvestment2020{FT};
            nTH::Int,
            siteids::Array{Int,1},
            selection::Array{String,1}
) where {FT<:AbstractFloat}

Run sensitivity analysis for optimal leaf investment, given

project LeafInvestment2020 type project

nTH Number of threads

siteids Indices of selected sites

selection List of parameters to do sensitivity on

Project NocturnalGS2020

ResearchProjects.plot_framework — Function

plot_framework(
            project::NocturnalGS2020{FT};
            saving::Bool
) where {FT<:AbstractFloat}

Plot model framework, given

project NocturnalGS2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_gswn_vs_pcd — Function

plot_gswn_vs_pcd(
            project::NocturnalGS2020{FT};
            saving::Bool
) where {FT<:AbstractFloat}

Plot gswn vs Psoil, Ca, and D, given

project NocturnalGS2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_gswn_vs_r — Function

plot_gswn_vs_r(
            project::NocturnalGS2020{FT};
            saving::Bool
) where {FT<:AbstractFloat}

Plot gswn vs leaf respiration rate, given

project NocturnalGS2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_gswn_vs_tem — Function

plot_gswn_vs_tem(
            project::NocturnalGS2020{FT};
            saving::Bool
) where {FT<:AbstractFloat}

Plot gswn vs temperature, given

project NocturnalGS2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_gswn_vs_time — Function

plot_gswn_vs_time(
            project::NocturnalGS2020{FT};
            saving::Bool
) where {FT<:AbstractFloat}

Plot gswn vs time, given

project NocturnalGS2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_gswn_vpd — Function

plot_gswn_vpd(
            project::NocturnalGS2020{FT};
            saving::Bool
) where {FT<:AbstractFloat}

Plot gswn vs VPD, given

project NocturnalGS2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.calculate_dAdE! — Function

calculate_dAdE!(
            leaf::Leaf{FT},
            env::AirLayer{FT},
            psm::AbstractPhotoModelParaSet{FT},
            a_n::FT,
            c_a::FT,
            e_leaf::FT,
            gbw::FT,
            p_atm::FT,
            par::FT,
            t_leaf::FT,
            vp_air::FT,
            dTdE::Bool
) where {FT<:AbstractFloat}

Calculate the marginal water use efficiency by firstly fitting the photosynthetic capacity and secondly numerically computing dAdE, given

leaf Leaf type struct that stores leaf photosynthetic parameters

env AirLayer type struct that stores environmental conditions

psm AbstractPhotoModelParaSet type struct that stores photosynthesis model parameters, like the temperature dependencies

a_n Given net photosynthetic rate

c_a Given CO₂ concentration in ppm

e_leaf Given leaf transpiration rate

gbw Given leaf boundary layer conductance for H₂O

p_atm Given atmospheric pressure

par Given photosynthetically active radiation

t_leaf Given leaf temperature

vp_air Given vapor pressure in the air

dTdE::Bool If true, account for leaf cooling from transpiration

ResearchProjects.calculate_dRdE — Function

calculate_dRdE(
            psm::AbstractPhotoModelParaSet{FT},
            r_n::FT,
            t_leaf::FT,
            wind::FT
) where {FT<:AbstractFloat}

Calculate marginal respiratory reduction, given

psm AbstractPhotoModelParaSet type struct that stores photosynthesis model parameters, like the temperature dependencies

r_n Given leaf respiration rate

t_leaf Given leaf temperature

wind Given wind speed

ResearchProjects.fitness_factor — Function

fitness_factor(
            project::NocturnalGS2020{FT};
            dTdE::Bool,
            display_result::Bool
) where {FT<:AbstractFloat}

Calculate the mean and standard deviation of the fitness factors, given

project NocturnalGS2020 type project control

dTdE Whether accounting for leaf cooling effect for daytime photosynthesis

display_result Optional. If true, display the results

Project OptimalGM2020

ResearchProjects.plot_data_fittings — Function

plot_data_fittings(
            proj::OptimalGM2020{FT};
            saving::Bool = false
) where {FT<:AbstractFloat}

Plot the model predictions and real data, given

proj OptimalGM2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_gm_var_maps — Function

plot_gm_var_maps(
            proj::OptimalGM2020{FT};
            saving::Bool = false
) where {FT<:AbstractFloat}

Plot the Ci-Gm and PAR-Gm contour map of ∂A/∂Gm, given

proj OptimalGM2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_optimal_gm_framework — Function

plot_optimal_gm_framework(
            proj::OptimalGM2020{FT};
            saving::Bool
) where {FT<:AbstractFloat}

Plot model framework, given

proj OptimalGM2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_optimal_gm_predictions — Function

plot_optimal_gm_predictions(
            proj::OptimalGM2020{FT};
            saving::Bool = false) where {FT<:AbstractFloat}

Plot model predictions of how optimal mesophyll respond to the environment, given

proj OptimalGM2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_optimal_gm_vs_ca — Function

plot_optimal_gm_vs_ca(
            proj::OptimalGM2020{FT};
            saving::Bool
) where {FT<:AbstractFloat}

Plot model framework, given

proj OptimalGM2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_optimal_gm_vs_gs — Function

plot_optimal_gm_vs_gs(
            proj::OptimalGM2020{FT};
            saving::Bool
) where {FT<:AbstractFloat}

Plot model framework, given

proj OptimalGM2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

ResearchProjects.plot_optimal_gm_vs_par — Function

plot_optimal_gm_vs_par(
            proj::OptimalGM2020{FT};
            saving::Bool
) where {FT<:AbstractFloat}

Plot model framework, given

proj OptimalGM2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font

SIFComparison2020

ResearchProjects.derive_spectrum — Function

derive_spectrum(proj::SIFComparison2020{FT};
                saving::Bool,
                use_latex::Bool) where {FT<:AbstractFloat}

Derive and plot the albedo and SIF spectrum, given

proj SIFComparison2020 type project control

saving Optional. If true, save the figure

use_latex Optional. If true, use latex and serif font